double to big

+++ src/main/java/com/transferwise/acorn/models/FeeAmount.java
@@ -1,6 +1,8 @@
 package com.transferwise.acorn.models;
 
+import java.math.BigDecimal;
+
 public class FeeAmount{
-    public double value;
+    public BigDecimal value;
     public String currency;
 }
